+++ src/logs/error.log
@@ -142,3 +142,6 @@
 2024-12-01 17:41:40 [ERROR]: File: CartController.js, line: 44, error: error is not defined
 2024-12-01 19:56:02 [ERROR]: File: CartController.js, line: 44, error: error is not defined
 2024-12-01 20:00:35 [ERROR]: File: CartController.js, line: 44, error: error is not defined
+2024-12-03 00:13:52 [ERROR]: File: CartController.js, line: 44, error: error is not defined
+2024-12-03 00:35:03 [ERROR]: File: CartController.js, line: 44, error: error is not defined
+2024-12-03 00:46:02 [ERROR]: File: CartController.js, line: 44, error: error is not defined
