+++ src/views/cart/cart.hbs
@@ -1,31 +1,4 @@
-
-<style type="text/css">
-body {
-  margin-top:20px;
-  background:#eee;
-}
-.ui-w-40 {
-    width: 40px !important;
-    height: auto;
-}
-
-.card{
-    box-shadow: 0 1px 15px 1px rgba(52,40,104,.08);    
-}
-
-.ui-product-color {
-    display: inline-block;
-    overflow: hidden;
-    margin: .144em;
-    width: .875rem;
-    height: .875rem;
-    border-radius: 10rem;
-    -webkit-box-shadow: 0 0 0 1px rgba(0,0,0,0.15) inset;
-    box-shadow: 0 0 0 1px rgba(0,0,0,0.15) inset;
-    vertical-align: middle;
-}
-</style>
-
+<link rel="stylesheet" href="/css/cart/cart.css">
 <body>
 <div class="container px-3 my-5 clearfix">
     <!-- Shopping cart table -->
@@ -57,7 +30,5 @@ body {
           </div>
       </div>
   </div>
-<script src="https://code.jquery.com/jquery-1.10.2.min.js"></script>
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@4.1.1/dist/js/bootstrap.bundle.min.js"></script>
 <script type="module" src="/js/cart/cart.js"></script>
 </body>
