+++ src/views/products/detail.hbs
@@ -216,4 +216,4 @@
             </div>
         </section>
 </div>
-<script src="/js/products/detail.js"></script>
+<script type="module" src="/js/products/detail.js"></script>
